# Conflicts:
#	app/src/main/java/com/applikey/mattermost/activities/ChannelDetailsActivity.java
#	app/src/main/java/com/applikey/mattermost/activities/DrawerActivity.java

+++ app/src/main/java/com/applikey/mattermost/activities/ChatActivity.java
@@ -12,7 +12,9 @@ import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
 import android.widget.EditText;
+import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
@@ -68,6 +70,18 @@ public class ChatActivity extends DrawerActivity implements ChatView {
     @Bind(R.id.et_message)
     EditText mEtMessage;
 
+    @Bind(R.id.ll_reply)
+    LinearLayout mLlReply;
+
+    @Bind(R.id.view_reply_separator)
+    View mViewReplySeparator;
+
+    @Bind(R.id.iv_reply_close)
+    ImageView mIvReplyClose;
+
+    @Bind(R.id.tv_reply_message)
+    TextView mTvReplyMessage;
+
     @InjectPresenter
     ChatPresenter mPresenter;
 
@@ -77,6 +91,8 @@ public class ChatActivity extends DrawerActivity implements ChatView {
 
     @Inject
     ImageLoader mImageLoader;
+
+    private String mRootId;
 
     private String mChannelId;
     private String mChannelName;
@@ -136,6 +152,7 @@ public class ChatActivity extends DrawerActivity implements ChatView {
         setToolbarText();
     }
 
+
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         menu.add(Menu.NONE, MENU_ITEM_SEARCH, Menu.NONE, R.string.search)
@@ -186,14 +203,24 @@ public class ChatActivity extends DrawerActivity implements ChatView {
 
     @OnClick(R.id.iv_send_message)
     public void onSend() {
-        mPresenter.sendMessage(mChannelId, mEtMessage.getText().toString());
+        if (mRootId == null) {
+            mPresenter.sendMessage(mChannelId, mEtMessage.getText().toString());
+        } else {
+            mPresenter.sendReplyMessage(mChannelId, mEtMessage.getText().toString(), mRootId);
+        }
     }
 
     @Override
     public void onMessageSent(long createdAt) {
-        mEtMessage.setText("");
+        mEtMessage.setText(null);
         mAdapter.setLastViewed(createdAt);
         scrollToStart();
+        hideReply();
+    }
+
+    @Override
+    public void showProgress(boolean enabled) {
+        mSrlChat.setRefreshing(enabled);
     }
 
     private void scrollToStart() {
@@ -210,6 +237,18 @@ public class ChatActivity extends DrawerActivity implements ChatView {
         mTvEmptyState.setVisibility(GONE);
     }
 
+    private void displayReply() {
+        mLlReply.setVisibility(VISIBLE);
+        mViewReplySeparator.setVisibility(VISIBLE);
+    }
+
+    private void hideReply() {
+        mLlReply.setVisibility(GONE);
+        mViewReplySeparator.setVisibility(GONE);
+        mTvReplyMessage.setText(null);
+        mRootId = null;
+    }
+
     private void setToolbarText() {
         final String prefix = !mChannelType.equals(Channel.ChannelType.DIRECT.getRepresentation())
                 ? CHANNEL_PREFIX : DIRECT_PREFIX;
@@ -217,25 +256,36 @@ public class ChatActivity extends DrawerActivity implements ChatView {
         mToolbar.setTitle(prefix + mChannelName);
     }
 
-    @Override
-    public void showProgress(boolean enabled) {
-        mSrlChat.setRefreshing(enabled);
-    }
-
-    private final PostAdapter.OnLongClickListener onPostLongClick = post -> {
-        final AlertDialog.Builder opinionDialogBuilder = new AlertDialog.Builder(this);
-        opinionDialogBuilder.setItems(R.array.post_own_opinion_array, (dialog, which) -> {
-            switch (which) {
-                case 0:
-                    deleteMessage(mChannelId, post);
-                    break;
-                case 1:
-                    editMessage(mChannelId, post);
-                    break;
-                default:
-                    throw new RuntimeException("Not implemented feature");
-            }
-        }).show();
+    private final PostAdapter.OnLongClickListener onPostLongClick = (post, isOwner) -> {
+        final AlertDialog.Builder dialogBuilder = new AlertDialog.Builder(this);
+        if (isOwner) {
+            dialogBuilder.setItems(R.array.post_own_opinion_array, (dialog, which) -> {
+                switch (which) {
+                    case 0:
+                        deleteMessage(mChannelId, post);
+                        break;
+                    case 1:
+                        editMessage(mChannelId, post);
+                        break;
+                    case 2:
+                        replyMessage(post);
+                        break;
+                    default:
+                        throw new RuntimeException("Not implemented feature");
+                }
+            });
+        } else {
+            dialogBuilder.setItems(R.array.post_opinion_array, (dialog, which) -> {
+                switch (which) {
+                    case 0:
+                        replyMessage(post);
+                        break;
+                    default:
+                        throw new RuntimeException("Not implemented feature");
+                }
+            });
+        }
+        dialogBuilder.show();
     };
 
     private void deleteMessage(String channelId, Post post) {
@@ -258,6 +308,12 @@ public class ChatActivity extends DrawerActivity implements ChatView {
                     mPresenter.editMessage(channelId, post, input.getText().toString());
                 })
                 .show();
+    }
+
+    private void replyMessage(Post post) {
+        displayReply();
+        mTvReplyMessage.setText(post.getMessage());
+        mRootId = post.getId();
     }
 
     private void initParameters() {
@@ -289,6 +345,8 @@ public class ChatActivity extends DrawerActivity implements ChatView {
                 hideKeyboard();
             }
         });
+
+        mIvReplyClose.setOnClickListener(v -> hideReply());
     }
 
     @Override
